==> Transnavigation/algoritm_Tol/Part0.Create_sites.ipynb <==
import pandas as pd
import geopandas as gpd
import numpy as np
import matplotlib.pyplot as plt
from datetime import datetime
import os
import math
import numpy as np
from sklearn.metrics.pairwise import haversine_distances

path='./data/'

# Считываем файл с последовательным списком остановок на маршрутах ('stop_seq.csv')
stop_seq = pd.read_csv(path + 'stop_seq_patp_for_28.csv', sep = ';', encoding = 'utf8')
stop_seq2 = pd.read_csv(path + 'stop_seq_TB.csv', sep = ';', encoding = 'utf8')

#stop_seq=stop_seq.append(stop_seq2)

stop_seq

stop_seq_gpd = gpd.GeoDataFrame(stop_seq,  geometry = [Point(xy) for xy in zip(stop_seq['ss_long'], stop_seq['ss_lat'])])
stop_seq_gpd.crs='epsg:4326'

#пересекли все остановки с остановками с буфером 300 метров
stops_with_buff = stop_seq_gpd.to_crs('epsg:32640')
stops_with_buff.geometry = stops_with_buff.geometry.buffer(100)
stops_with_buff = stops_with_buff.to_crs('epsg:4326')
stops_new=gpd.sjoin(stop_seq_gpd, stops_with_buff, how='left', op='within', lsuffix='left', rsuffix='right')

stops_pares=stops_new[stops_new['mr_id_left']==stops_new['mr_id_right']].reset_index()

stops_pares=stops_pares[['ssId_left','ss_lat_left','ss_long_left','ssId_right','ss_lat_right','ss_long_right']].drop_duplicates()

#stops_pares=stops_pares[stops_pares['ssId_left']!=stops_pares['ssId_right']]

stops_pares['geometry']=Point(zip(stops_pares['ss_long_right'], stops_pares['ss_lat_right']))

stops_pares_gpd = gpd.GeoDataFrame(stops_pares,  geometry = 'geometry')
stops_pares_gpd.crs='epsg:4326'

stops_pares_gpd_group=stops_pares_gpd.dissolve(by='ssId_left')

stops_pares

stops_pares_group=stops_pares.groupby(['ssId_left','ss_lat_left','ss_long_left'])['ssId_right'].apply(list).reset_index()
stops_pares_group2=stops_pares.groupby(['ssId_left','ss_lat_left','ss_long_left'])['ss_lat_right'].apply(list).reset_index()
stops_pares_group3=stops_pares.groupby(['ssId_left','ss_lat_left','ss_long_left'])['ss_long_right'].apply(list).reset_index()

stops_pares_group4=stops_pares.groupby(['ssId_left','ss_lat_left','ss_long_left']).agg({'ss_lat_right': np.mean}).reset_index().rename(columns={'ss_lat_right':'ss_lat'})
stops_pares_group5=stops_pares.groupby(['ssId_left','ss_lat_left','ss_long_left']).agg({'ss_long_right': np.mean}).reset_index().rename(columns={'ss_long_right':'ss_long'})

stops_pares_group=stops_pares_group.merge(stops_pares_group2,on=['ssId_left','ss_lat_left','ss_long_left'],how='left')\
.merge(stops_pares_group3,on=['ssId_left','ss_lat_left','ss_long_left'],how='left')\
.merge(stops_pares_group4,on=['ssId_left','ss_lat_left','ss_long_left'],how='left')\
.merge(stops_pares_group5,on=['ssId_left','ss_lat_left','ss_long_left'],how='left')

sites=pd.DataFrame(stops_pares_group[['ss_lat','ss_long']].drop_duplicates().reset_index())

sites=sites.rename(columns={'index':'site_id'})

stops_pares_group=stops_pares_group.merge(sites,on=['ss_lat','ss_long'],how='left')

stops_pares_group.to_csv('stops_sites_DOP.csv')

len(stops_pares_group)

stops_pares_group=pd.read_csv('stops_sites_coordinates.csv')

stops_pares_group=stops_pares_group[['ssId_left','ss_lat_left','ss_long_left','site_id']]

site_group=stops_pares_group.groupby(['site_id']).agg({'ss_lat_left': np.mean,'ss_long_left': np.mean}).reset_index().rename(columns={'ss_lat_left':'ss_lat','ss_long_left':'ss_long'})

site_group['site_id_old']=site_group['site_id']

site_group['site_id']=site_group.index

site_group.to_csv('sites_coordinate.csv', index=False)

stops_pares_group=stops_pares_group.merge(site_group, left_on='site_id', right_on='site_id_old')

stops_pares_group=stops_pares_group[['ssId_left','ss_lat','ss_long','site_id_y']]

stops_pares_group

stop_seq=stop_seq.merge(stops_pares_group[['ssId_left','site_id_y']], left_on='ssId',right_on='ssId_left')

stop_seq=stop_seq[['ssId','rc_orderby','ss_lat','ss_long','rl_id','mr_id','site_id_y']]

stop_seq=stop_seq.rename(columns={'site_id_y':'site_id'})

stop_seq=stop_seq.sort_values(by=['mr_id','rl_id','rc_orderby'])

stop_seq.to_csv('stop_seq.csv',index=False)
