==> Transnavigation/algoritm_Tol/Part2.Join_with_validation_TB.ipynb <==
import pandas as pd
import geopandas as gpd
import numpy as np
import matplotlib.pyplot as plt
from datetime import datetime
import os
from tqdm import tqdm_notebook as tqdm

path = '.\\valid\\'

path2 = '.\\res\\'

Считываем данные по трекам с привязанными остановками

## НЕ ЗАБЫТЬ ПОМЕНЯТЬ ИМЯ ФАЙЛА

res_stop = pd.read_csv(path2+'res_TB_2019-10-09.csv')

res_stop=res_stop.rename(columns={'MODULEID':'tabelnum'})

Определяем времена начала и конца для каждого табельного номера

mr_id_marsh_valid_id = pd.read_csv('./data/tol_marsh_TB.csv', sep = ';', encoding = 'windows-1251')

mr_id_marsh_valid_id.head()

validation= pd.read_csv(path+'2019.10.09.csv', sep = ',', encoding = 'cp1251')
validation=validation.merge(mr_id_marsh_valid_id,on =['Код Маршрута'])
validation=validation[validation['mr_id'].isin(res_stop['mr_id'].unique())]
validation['time']=pd.to_datetime(validation['Дата и время поездки'])
validation['hour']=validation['time'].dt.hour
validation.loc[validation.hour ==0, 'hour'] = 24
validation.loc[validation.hour ==1, 'hour'] = 25
validation.loc[validation.hour ==2, 'hour'] = 26
validation.loc[validation.hour ==3, 'hour'] = 27
validation['min']=validation['time'].dt.minute
validation['sec']=validation['time'].dt.second
validation['time']=(validation['hour'].astype(int))*3600+(validation['min'].astype(int))*60+(validation['sec'].astype(int))
validation=validation.drop(['hour'], axis=1).drop(['min'], axis=1).drop(['sec'], axis=1)
validation_start=pd.DataFrame(validation.groupby(['Код Маршрута','Табельный номер']).agg({'time': np.min})).reset_index().rename(columns={"time": "start"})
validation_finish=pd.DataFrame(validation.groupby(['Код Маршрута','Табельный номер']).agg({'time': np.max})).reset_index().rename(columns={"time": "finish"})
validation_start_finish=validation_start.merge(validation_finish, on=['Код Маршрута','Табельный номер'])
validation2=validation.merge(validation_start_finish, on=['Код Маршрута','Табельный номер'])
del validation_start
del validation_finish
del validation

## НЕ ЗАБЫТЬ ПОМЕНЯТЬ ИМЯ ФАЙЛА

df_4 = pd.read_csv('./data/TB_2019-10-09.csv')
df_4=df_4.rename(columns={'MODULEID':'tabelnum'})
df_4['time2']=pd.to_datetime(df_4['time'])
df_4['hour']=df_4['time2'].dt.hour
df_4['min']=df_4['time2'].dt.minute
df_4['sec']=df_4['time2'].dt.second
df_4.loc[df_4.hour ==0, 'hour'] = 24
df_4.loc[df_4.hour ==1, 'hour'] = 25
df_4.loc[df_4.hour ==2, 'hour'] = 26
df_4.loc[df_4.hour ==3, 'hour'] = 27
df_4['time2']=(df_4['hour'].astype(int))*3600+(df_4['min'].astype(int))*60+(df_4['sec'].astype(int))
df_4=df_4.drop(['hour'], axis=1).drop(['min'], axis=1).drop(['sec'], axis=1)
df_4_start=pd.DataFrame(df_4.groupby(['mr_id','tabelnum']).agg({'time2': np.min})).reset_index().rename(columns={"time2": "start"})
df_4_finish=pd.DataFrame(df_4.groupby(['mr_id','tabelnum']).agg({'time2': np.max})).reset_index().rename(columns={"time2": "finish"})
df_4_start_finish=df_4_start.merge(df_4_finish, on=['mr_id','tabelnum']).sort_values(by=['mr_id', 'start'])
del df_4_start
del df_4_finish

validation_start_finish2=validation_start_finish.merge(mr_id_marsh_valid_id,on =['Код Маршрута'])
mrid=validation_start_finish2['mr_id'].unique()
df_4_start_finish=df_4_start_finish[df_4_start_finish['mr_id'].isin(mrid)]
validation_start_finish2=validation_start_finish2.sort_values(by=['mr_id', 'start'])
df_4_start_finish=df_4_start_finish.sort_values(by=['mr_id', 'start'])

validation_start_finish2.head()

df_4_start_finish_mem=df_4_start_finish

res=validation_start_finish2

1. Раскидываем машины по табельным номерам валидаций. Жесткое условие (все валидации внутри трека)

k=0
res=validation_start_finish2
tabelnum = []
start = []
finish = []
tracks=pd.DataFrame()
for index, row in res.iterrows():
    stops=df_4_start_finish[(df_4_start_finish['mr_id']==row['mr_id'])
                    &(df_4_start_finish['start']<=row['start'])
                    &(df_4_start_finish['finish']>row['finish'])]
    df_4_start_finish=df_4_start_finish[~df_4_start_finish.isin(stops[0:1])].dropna(how = 'all')
    if stops.empty:
        tabelnum.append(0)
        start.append(0)
        finish.append(0)
    else:
        tabelnum_unic=stops['tabelnum'].values
        tabelnum.append(tabelnum_unic[0])
        start_unic=stops['start'].values
        start.append(start_unic[0])
        finish_unic=stops['finish'].values
        finish.append(finish_unic[0])
res['tabelnum']=tabelnum
res['start_track']=start
res['finish_track']=finish

res.head()

len(df_4_start_finish)

2. Раскидываем машины по табельным номерам валидаций. Мягкое условие 1 (все валидации после начала трека)

k=0
tabelnum = []
start = []
finish = []
stops=pd.DataFrame()
for index, row in res.iterrows():
    if row['tabelnum']==0:
        stops=df_4_start_finish[(df_4_start_finish['mr_id']==row['mr_id'])
                    &(df_4_start_finish['start']<=row['start'])]
        df_4_start_finish=df_4_start_finish[~df_4_start_finish.isin(stops[0:1])].dropna(how = 'all')
        if stops.empty:
            tabelnum.append(0)
            start.append(0)
            finish.append(0)
        else:
            tabelnum_unic=stops['tabelnum'].values
            tabelnum.append(tabelnum_unic[0])
            start_unic=stops['start'].values
            start.append(start_unic[0])
            finish_unic=stops['finish'].values
            finish.append(finish_unic[0])
    else:
        tabelnum.append(0)
        start.append(0)
        finish.append(0)        
res['tabelnum_2']=tabelnum
res['start_track_2']=start
res['finish_track_2']=finish

3. Раскидываем машины по табельным номерам валидаций. Мягкое условие 2 (все валидации до конца трека)

k=0
tabelnum = []
start = []
finish = []
stops=pd.DataFrame()
for index, row in res.iterrows():
    if row['tabelnum']==0:
        stops=df_4_start_finish[(df_4_start_finish['mr_id']==row['mr_id'])
                    &(df_4_start_finish['finish']>row['finish'])]
        df_4_start_finish=df_4_start_finish[~df_4_start_finish.isin(stops[0:1])].dropna(how = 'all')
        if stops.empty:
            tabelnum.append(0)
            start.append(0)
            finish.append(0)
        else:
            tabelnum_unic=stops['tabelnum'].values
            tabelnum.append(tabelnum_unic[0])
            start_unic=stops['start'].values
            start.append(start_unic[0])
            finish_unic=stops['finish'].values
            finish.append(finish_unic[0])
    else:
        tabelnum.append(0)
        start.append(0)
        finish.append(0)        
res['tabelnum_3']=tabelnum
res['start_track_3']=start
res['finish_track_3']=finish

res.head()

len(df_4_start_finish)

4.Дозаполняем табельные номера тем, что не вошло

k=0
df_4_start_finish=df_4_start_finish_mem
tabelnum = []
start = []
finish = []
stops=pd.DataFrame()
for index, row in res.iterrows():
    if row['tabelnum']==0:
        stops=df_4_start_finish[(df_4_start_finish['mr_id']==row['mr_id'])
                    &(df_4_start_finish['finish']>row['finish'])]
        if stops.empty:
            tabelnum.append(0)
            start.append(0)
            finish.append(0)
        else:
            tabelnum_unic=stops['tabelnum'].values
            tabelnum.append(tabelnum_unic[0])
            start_unic=stops['start'].values
            start.append(start_unic[0])
            finish_unic=stops['finish'].values
            finish.append(finish_unic[0])
    else:
        tabelnum.append(0)
        start.append(0)
        finish.append(0)        
res['tabelnum_4']=tabelnum
res['start_track_4']=start
res['finish_track_4']=finish

res.to_csv(path+'for_mege_tabelnum_2019-10-09.csv')
#df_4_start_finish.to_csv(path+'for_mege_tabelnum_ost_29.csv')

len(res[res['tabelnum_3']>0])

len(res[res['tabelnum']==0])

Перезаписываем данные в столбцах

res.loc[res.tabelnum==0, 'tabelnum'] = res.loc[res.tabelnum==0, 'tabelnum_3']
res.loc[res.tabelnum==0, 'tabelnum'] = res.loc[res.tabelnum==0, 'tabelnum_2']
res.loc[res.start_track==0, 'start_track'] = res.loc[res.start_track==0, 'start_track_2']
res.loc[res.start_track==0, 'start_track'] = res.loc[res.start_track==0, 'start_track_3']
res.loc[res.finish_track==0, 'finish_track'] = res.loc[res.finish_track==0, 'finish_track_2']
res.loc[res.finish_track==0, 'finish_track'] = res.loc[res.finish_track==0, 'finish_track_3']

res.loc[res.tabelnum==0, 'tabelnum'] = res.loc[res.tabelnum==0, 'tabelnum_4']
res.loc[res.start_track==0, 'start_track'] = res.loc[res.start_track==0, 'start_track_4']
res.loc[res.finish_track==0, 'finish_track'] = res.loc[res.finish_track==0, 'finish_track_4']

Пересохранили табельный номер как int, чтобы правильнее можно было merge делать

table_num_merge_29=res[['Табельный номер','mr_id','tabelnum']]

len(validation2['mr_id'].unique())

table_num_merge_29[['Табельный номер','tabelnum']]

table_num_merge_29.info()

validation4=validation2.merge(table_num_merge_29,on =['mr_id','Табельный номер'],how='left')
validation4=validation4.sort_values(by=['mr_id','Табельный номер','time'])

validation4

del validation2

Переводим столбец с временем в формат INT и прописываем для каждого табельного номера время начала и конца

res_stop

validation4['tr_id']='';

validation4=validation4[validation4['mr_id'].isin(res_stop['mr_id'].unique())]

Смотрим сколько маршрутов у нас есть и в валидациях, и в треках

len(validation4['mr_id'].unique())

validation4['tabelnum'].unique()

res_stop2=res_stop

k=0
res=validation4
result = []
result2 = []
stops=pd.DataFrame()
for index, row in validation4.iterrows():
    stops=res_stop2[(res_stop2['mr_id']==row['mr_id'])
                    &(res_stop2['tabelnum']==row['tabelnum'])
                    &(res_stop2['time_start']<=row['time'])
                    &((res_stop2['time_finish']>row['time'])|(res_stop2['time_finish'].isna()))]
    if stops.empty:
        result.append(0)
        result2.append(0)
    else:
        stop_unic=stops['ssId'].values
        result.append(stop_unic[0])
        rl_id=stops['rl_id'].values
        result2.append(rl_id[0])
res['ssId']=result
res['rl_id']=result2
FOR_MATRIX=res

len(res[res['ssId']==0])

stops=pd.read_csv('./data/stop_seq_TB.csv', sep = ';', encoding='utf-8')

stops=stops[['ssId','ss_lat','ss_long']].drop_duplicates()

stops

res2=res.merge(stops, on='ssId', how='left')

res2['c']=1

res2[res2['ssId']==0]['mr_id'].unique()

rl_orderby=res_stop[['ssId','rl_id','mr_id','rc_orderby']].drop_duplicates()

rl_orderby.head()

len(res2)

res2=res2.merge(rl_orderby, on=['ssId','rl_id','mr_id'], how='left')

80874/288226

len(res2[res2['ssId']==0])

## НЕ ЗАБЫТЬ ПОМЕНЯТЬ ИМЯ ФАЙЛА

res2.to_csv(path+'For_matrix_TB_2019-10-09.csv')

res2.head()

for_karta3=res2.groupby(['ssId','ss_lat','ss_long']).agg({'c': np.sum}).reset_index()
for_karta3=for_karta3.groupby(['ssId','ss_lat','ss_long']).agg({'c': np.sum}).reset_index()
gdf = gpd.GeoDataFrame(
    for_karta3, geometry=gpd.points_from_xy(x=for_karta3.ss_long, y=for_karta3.ss_lat)
)
gdf.crs='epsg:4326'
gdf.to_file(fr"./res/stops_2019_10_09.json", index=False, driver="GeoJSON")
