==> Transnavigation/algoritm_Tol/Part5.Matrix_HOUR_PICK.ipynb <==
import pandas as pd
import geopandas as gpd
import numpy as np
import matplotlib.pyplot as plt
from datetime import datetime
import os
import math
from tqdm import tqdm_notebook as tqdm
import warnings
warnings.filterwarnings('ignore')
from shapely.geometry import LineString
from shapely.geometry import Point
from math import radians, cos, sin, asin, sqrt

path=('./valid/')

DATA=pd.read_csv(path+'For_matrix_TB_2019-10-07.csv', sep = ',', encoding='utf-8')
DATA2=pd.read_csv(path+'For_matrix_TB_2019-10-08.csv', sep = ',', encoding='utf-8')
DATA2['time']=DATA2['time']+24*3600 #прибалвяем 24 часа в секундах
DATA3=pd.read_csv(path+'For_matrix_TB_2019-10-09.csv', sep = ',', encoding='utf-8')
DATA3['time']=DATA3['time']+24*3600+24*3600
DATA=DATA.append(DATA2).append(DATA3)

DATA=DATA[['Дата и время поездки','Номер кристалла','TCAT','Код Маршрута','Код предприятия_x','mr_id','time','ssId','rl_id','rc_orderby']]

len(DATA)

del DATA2
del DATA3

Привязали все валидации к сайтам

sites=pd.read_csv('stop_seq.csv', sep = ',', encoding='utf-8')

sites[0:2]

sites=sites[['ssId','site_id']].drop_duplicates()
DATA=DATA.merge(sites,on='ssId', how='left')

DATA[0:2]

ВЫБРАЛИ ДАННЫЕ ЗА ЧАС ПИК

#выбрала час пик только из перовго дня
DATA_pick=DATA[DATA['Дата и время поездки'].str.contains( '2019-10-07')] 
DATA_pick=DATA_pick[DATA_pick['Дата и время поездки'].str.contains( ' 06:')]\
    .append(DATA_pick[DATA_pick['Дата и время поездки'].str.contains( ' 07:')])\
    .append(DATA_pick[DATA_pick['Дата и время поездки'].str.contains( ' 08:')])

len(DATA_pick)

Убираем и запоминаем из выборки тех, у кого нет номера кристалла (невозможно точно привязать точку выхода)

DATA_bank_pick=DATA_pick[DATA_pick['Номер кристалла'].isna()]

ОСТАВЛЯЕМ В ЧАСЕ ПИК ТОЛЬКО ТЕХ, у кого есть номер кристалла:

DATA_pick=DATA_pick[~DATA_pick.index.isin(DATA_bank_pick.index)]

DATA_pick['time'].min()

DATA_pick['time'].max()

Выбрала из общего массива только те данные, где встречаются валидации из пикового промежутка:

DATA=DATA[DATA['Номер кристалла'].isin(DATA_pick['Номер кристалла'].unique())]

len(DATA)

Выбрала только те значения, которые встречаются спустя 3 часа после начала часа пик

DATA=DATA[DATA['time']>DATA_pick['time'].min()+3*3600]

DATA=DATA.sort_values(['Номер кристалла','time','site_id'])

DATA_pick=DATA_pick.sort_values(['Номер кристалла','time','site_id'])

len(DATA_pick)

len(DATA)

from tqdm import tqdm
k=0
res=DATA_pick
result = []
result2 = []
result3 = []
result4 = []
stops=pd.DataFrame()
with tqdm(total=DATA_pick.shape[0]) as pbar: 
    for index, row in DATA_pick.iterrows():
        pbar.update(1)
        stops=DATA[(DATA['Номер кристалла']==row['Номер кристалла'])
                    &(DATA['time']>(row['time']+3600*3))
                    &(DATA['site_id']!=row['site_id'])]
        if stops.empty:
            result.append(0)
            result2.append(0)
            result3.append(0)
            result4.append(0)
        else:
            stop_unic=stops['site_id'].values
            result.append(stop_unic[0])
            time_2=stops['time'].values
            result2.append(time_2[0])
            rl_id=stops['rl_id'].values
            result3.append(rl_id[0])
            rc_orderby=stops['rc_orderby'].values
            result4.append(rc_orderby[0])
res['site_id_2']=result
res['time_2']=result2
res['rl_id_2']=result3
res['rc_orderby_2']=result4

res.to_csv('res_2019_10_07.csv')

#res=pd.read_csv('res_2019_10_16.csv')

len(res)

len(DATA_bank_pick)

len(res[(res['site_id_2']>0)&(res['site_id']>0)&(res['site_id_2']!=res['ssId'])])

len(res[(res['site_id_2']>0)&(res['site_id']>0)&(res['site_id_2']!=res['ssId'])])/len(res)

res=res.fillna(0)

res['c']=1

res['c'].sum()

res=res.groupby(['site_id','site_id_2','mr_id']).agg({'c':np.nansum}).reset_index() 

res['c'].sum()

Заполняем неприязавшиеся Site_ID наиболее подходящими ID сайтов на маршруте

res_not_null=res[(res['site_id_2']>0)&(res['site_id']>0)] #выбрали из матрицы те строки, где все значения есть
res_null_last_stops=res[((res['site_id_2']==0)|(res['site_id_2'].isna()))
                        &(res['site_id']>0)].drop(['site_id_2'], axis=1) #выбрали только те строки, где есть остановка входа, но нет остановки выхода
res_null_first_stops=res[((res['site_id']==0)|(res['site_id'].isna()))
                        &(res['site_id_2']>0)].drop(['site_id'], axis=1) #выбрали только те строки, где есть остановка вЫхода, но нет остановки входа

Записываем всё, что не записалось и не определилось в массив res_null_null

res_null_null=res[~res.index.isin(res_not_null.index)]
res_null_null2=res_null_null[~res_null_null.index.isin(res_null_last_stops.index)]
del res_null_null
res_null_null=res_null_null2[~res_null_null2.index.isin(res_null_first_stops.index)].drop(['site_id'], axis=1).drop(['site_id_2'], axis=1)
del res_null_null2


res_null_null=res_null_null.reset_index()
res_null_last_stops=res_null_last_stops.reset_index()
res_null_first_stops=res_null_first_stops.reset_index()

len(res)-(len(res_not_null)+len(res_null_null)+len(res_null_last_stops)+len(res_null_first_stops))

res['c'].dropna().sum()

res_null_null['c'].unique()

Добавляем банк

#добавили в выборку, где не привязались посл.остановки баноквские карты, где есть валидация начала
DATA_bank_ssid=DATA_bank_pick[DATA_bank_pick['site_id']>0]
DATA_bank_ssid['c']=1
DATA_bank_ssid=DATA_bank_ssid[['mr_id','site_id','c']]
res_null_last_stops=res_null_last_stops.append(DATA_bank_ssid).reset_index()

#добавили в выборку, где ничего не привязалось баноквские карты, где нет валидаций начала
DATA_bank_null=DATA_bank_pick[~DATA_bank_pick.index.isin(DATA_bank_ssid.index)]
DATA_bank_null['c']=1
DATA_bank_null=DATA_bank_null[['mr_id','site_id','c']]
res_null_null=res_null_null.append(DATA_bank_null).reset_index()

len(res)+len(DATA_bank_pick)-(len(res_not_null)+len(res_null_null)+len(res_null_last_stops)+len(res_null_first_stops))

Сгруппировали все выгрузки, чтобы меньше хранить в оперативной памяти:

res_null_null=res_null_null.fillna(0)

res_null_null=res_null_null.groupby(['mr_id']).agg({'c':np.sum}).reset_index() 

res_null_last_stops=res_null_last_stops.groupby(['mr_id','site_id']).agg({'c':np.sum}).reset_index() 

res_null_first_stops=res_null_first_stops.groupby(['mr_id','site_id_2']).agg({'c':np.sum}).reset_index() 

res_not_null=res_not_null.groupby(['mr_id','site_id','site_id_2']).agg({'c':np.sum}).reset_index() 

res_not_null=res_not_null.sort_values(by=['c'],ascending=False)

res_null_null['c'].sum()+res_null_last_stops['c'].sum()+res_null_first_stops['c'].sum()+res_not_null['c'].sum()

res_null_last_stops['c'].sum()

res_not_null['c'].sum()

Определяем наиболее подходящие остановки для высадки на маршруте:

#Размазываем вход-выход для случая, когда мы знаем ВХОД
def input_func(res_not_null):
    input_=res_not_null.groupby(['site_id','site_id_2']).agg({'c':np.sum}).reset_index() 
    sum_mr_id=input_.groupby(['site_id']).agg({'c':np.sum}).reset_index() 
    input_=input_.merge(sum_mr_id, on =['site_id'])
    input_['c_x']=input_['c_x']/input_['c_y']
    input_=input_.drop(['c_y'], axis=1).rename(columns={'c_x':'c'})
    return input_

#Размазываем вход-выход для случая, когда мы знаем ВЫХОД
def exit_func(res_not_null):
    exit=res_not_null.groupby(['site_id','site_id_2']).agg({'c':np.sum}).reset_index() 
    sum_mr_id=exit.groupby(['site_id_2']).agg({'c':np.sum}).reset_index() 
    exit=exit.merge(sum_mr_id, on =['site_id_2'])
    exit['c_x']=exit['c_x']/exit['c_y']
    exit=exit.drop(['c_y'], axis=1).rename(columns={'c_x':'c'})
    return exit

#Там где не определился конец поездки, выбираем наиболее подходящую SsID для выхода (пропорционально выходу)
input_=input_func(res_not_null)
res_null_last_stops2=res_null_last_stops.merge(input_,on=['site_id'],how='left')
res_null_last_stops2['c']=res_null_last_stops2['c_x']*res_null_last_stops2['c_y']
res_null_last_stops2=res_null_last_stops2.groupby(['mr_id','site_id','site_id_2']).agg({'c':np.sum}).reset_index()

#записываем результат
res_null_last_stops2=res_null_last_stops2.fillna(0)
res_not_null=res_not_null.append(res_null_last_stops2[res_null_last_stops2['site_id_2']>0])

#записываем оставшиеся непривязавшиеся остановки
res_null_last_stops3=res_null_last_stops2[res_null_last_stops2['site_id_2']==0]
res_null_null=res_null_null.append(res_null_last_stops3.drop(['site_id_2'], axis=1).drop(['site_id'], axis=1))
del res_null_last_stops2
del res_null_last_stops

#Там где не определился конец поездки, выбираем наиболее подходящую SsID для выхода (пропорционально входу)
exit=exit_func(res_not_null)
res_null_first_stops2=res_null_first_stops.merge(exit,on=['site_id_2'],how='left')
res_null_first_stops2['c']=res_null_first_stops2['c_x']*res_null_first_stops2['c_y']
res_null_first_stops2=res_null_first_stops2.groupby(['mr_id','site_id','site_id_2']).agg({'c':np.sum}).reset_index()
res_null_first_stops2=res_null_first_stops2.fillna(0)
#записываем результат
res_not_null=res_not_null.append(res_null_first_stops2[res_null_first_stops2['site_id_2']>0])

#записываем оставшиеся непривязавшиеся остановки
res_null_first_stops3=res_null_first_stops2[res_null_first_stops2['site_id_2']==0]
res_null_null=res_null_null.append(res_null_first_stops3.drop(['site_id_2'], axis=1).drop(['site_id'], axis=1))
del res_null_first_stops2
del res_null_first_stops
del res_null_first_stops3

Снова группируем

res_not_null=res_not_null.groupby(['mr_id','site_id','site_id_2']).agg({'c':np.sum}).reset_index() 

res_not_null['c'].sum()

#Размазываем вход-выход для случая, когда мы НИЧЕГО НЕ ЗНАЕМ
exit_input=res_not_null.groupby(['mr_id','site_id','site_id_2']).agg({'c':np.sum}).reset_index() 
sum_mr_id=exit_input.groupby(['mr_id']).agg({'c':np.sum}).reset_index() 
exit_input=exit_input.merge(sum_mr_id, on ='mr_id')
exit_input['c_x']=exit_input['c_x']/exit_input['c_y']
exit_input=exit_input.drop(['c_y'], axis=1).rename(columns={'c_x':'c'})

exit_input

#Там где ничего не определилось, размазываем валидацию по всем парам вероятным :
res_null_null2=res_null_null.merge(exit_input,on=['mr_id'],how='left')

res_null_null2

res_null_null2['c']=res_null_null2['c_x']*res_null_null2['c_y']

res_null_null2=res_null_null2.drop('c_x',axis=1).drop('c_y',axis=1)

res_not_null['c'].sum()

res_not_null2=res_not_null.append(res_null_null2)

res_null_null2=res_null_null2.fillna(0)

res_not_null2=res_not_null2.groupby(['site_id','site_id_2']).agg({'c':np.sum}).reset_index()

itog=res_not_null2['c'].sum()

itog

#DAY=pd.read_csv(path+'For_matrix_2019_10_16.csv', sep = ',', encoding='utf-8')

Кол-во валидаций на НГПТ за день составило:

#day=len(DAY)

#day

Кол-во валидаций на НГПТ за час пик день составило:

#DAY_pick=DAY[DAY['Дата и время поездки'].str.contains( ' 07:')]\
#    .append(DAY[DAY['Дата и время поездки'].str.contains( ' 08:')])\
#    .append(DAY[DAY['Дата и время поездки'].str.contains( ' 09:')])
#pick=len(DAY_pick)

#pick
#Годовой поток по пригородным маршрутам составил 15369619
#Переход к месяцу Октябрь от года - 10,73 (делим)
#Переход к одному дню октября = 396 889 (средний будний день в октябре 2019 года)/10 852 053 (суммарный месячный поток в октябре 2019 за год)
#Переход к часу - пик - делим кол-во валидаций за день к количеству валидаций за час пик

#res_not_null2['c']=res_not_null2['c']+(15369619/10.73)*(396889/10852053)*(pick/day)/itog

#res_not_null2['c'].sum()

matrix=res_not_null2.groupby(['site_id','site_id_2']).agg({'c':np.sum}).reset_index()

matrix=matrix.sort_values(by='c',ascending=False)

matrix['site_id']=matrix['site_id'].astype('int64')

matrix['c']=round(matrix['c'])

matrix['c']=matrix['c'].astype(int)

matrix['site_id_2']=matrix['site_id_2'].astype('int64')

matrix['c'].sum()

matrix.to_csv('./res/matrix_oct_TB.csv', index=False)

matrix2=pd.read_csv('./res/matrix_sep_patp.csv')

matrix_all=matrix.append(matrix2)

matrix_all=pd.DataFrame(matrix_all.groupby(['site_id','site_id_2']).agg({'c':np.sum}).reset_index())

matrix_all.to_csv('./res/matrix_Tolyatti.csv', index=False)

## Блок проверок и визуализации

## ДОБАВЛЯЕМ ВИДЕОАНАЛИТИКУ

#считали данные видеоаналитики
video_data=pd.read_csv('Videoanal.csv',encoding='utf-8', sep=";")
video_data=video_data[video_data['Широта'].notna()]
video_data=video_data.fillna(0)
video_data['Открытие двери']=video_data['Открытие двери'].astype(str)


video_data1=video_data[video_data['Выход на маршрут'].str.startswith('06')].\
append(video_data[video_data['Выход на маршрут'].str.startswith('07')]).\
append(video_data[video_data['Выход на маршрут'].str.startswith('08')])

video_data1=video_data1.merge(sites,left_on='ID остановки (Яндекс)',right_on='ssId',how='left')

#перешли к входу среднему на одном рейсе
video_data2=pd.DataFrame(video_data1.groupby(['route_id','trip_id','ID остановки (Яндекс)','site_id']).agg({'Вошло':np.mean}).reset_index())

#cчитали частоту хождения по маршрутам
freq=pd.read_csv('freq2.csv',encoding='utf-8',sep=';')

freq=freq.fillna(0)

freq[0:2]

#перешли к входу среднему на одном рейсе
video_data2=pd.DataFrame(video_data2.groupby(['trip_id','site_id']).agg({'Вошло':np.mean}).reset_index())

video_data2[0:2]

#перешли к среднему выходу на одном рейсе автобусов 
vyxod_reis=pd.DataFrame(video_data1.groupby(['route_id','trip_id','ID остановки (Яндекс)','site_id']).agg({'Вышло':np.mean}).reset_index())

vyxod_reis=pd.DataFrame(vyxod_reis.groupby(['trip_id','site_id']).agg({'Вышло':np.sum}).reset_index())

#выбрали только те остановки,по которым будем размазывать выход (убрали те остановки, на которых у нас вход и там, где выход = 0)
#vyxod_reis=vyxod_reis[~vyxod_reis['site'].isin(video_data2['site'].unique())]
vyxod_reis=vyxod_reis[vyxod_reis['Вышло']>0]

#считаем суммарный выход на одном направлении на одном маршруте:
vyxod_reis_group_trip_id=pd.DataFrame(vyxod_reis.groupby(['trip_id']).agg({'Вышло':np.sum}).reset_index())

#размазываем 
vyxod_reis=vyxod_reis.merge(vyxod_reis_group_trip_id, on ='trip_id',how='outer')
vyxod_reis['part']=vyxod_reis['Вышло_x']/vyxod_reis['Вышло_y']
vyxod_reis=vyxod_reis[['trip_id','site_id','part']]

matrix_video=video_data2.merge(vyxod_reis, on ='trip_id', how='outer')

matrix_video['c']=matrix_video['Вошло']*matrix_video['part']

matrix_video=pd.DataFrame(matrix_video.groupby(['site_id_x','site_id_y']).agg({'c':np.sum}).reset_index())

matrix_video=matrix_video[matrix_video['site_id_x']!=matrix_video['site_id_y']]

matrix_video[0:2]

matrix_video=matrix_video.rename(columns={'site_id_x':'site_id','site_id_y':'site_id_2'})

freq[0:2]

matrix_video=matrix_video.merge(freq,on ='site_id', how='left')

matrix_video['c']=matrix_video['c']*matrix_video['freq']

matrix_video['c'].sum()

matrix_all['c'].sum()

matrix_all[0:1]

matrix_video[0:1]

matrix_video=matrix_video[['site_id','site_id_2','c']]

matrix_all=matrix_all.append(matrix_video)

matrix_all_with_video=pd.DataFrame(matrix_all.groupby(['site_id','site_id_2']).agg({'c':np.sum}).reset_index())

matrix_all_with_video['c'].sum()

matrix_all_with_video['c']=np.ceil(matrix_all_with_video['c'])

matrix_all_with_video['c']=matrix_all_with_video['c'].astype('int64')
matrix_all_with_video['site_id']=matrix_all_with_video['site_id'].astype('int64')
matrix_all_with_video['site_id_2']=matrix_all_with_video['site_id_2'].astype('int64')

matrix_all_with_video['c'].sum()

matrix_all_with_video=matrix_all_with_video[matrix_all_with_video['c']>0]

matrix_all_with_video.to_csv('./res/matrix_with_video.csv')

### Блок проверок и визуализации

site=pd.read_csv('sites_coordinate.csv',sep=';')
matrix_all_s=matrix_all_with_video.merge(site,on='site_id',how='left')
matrix_all_s=matrix_all_s.merge(site,left_on='site_id_2',right_on='site_id',how='left')
matrix_all_s=matrix_all_s[['site_id_x','site_id_2','c','ss_lat_x','ss_long_x','ss_lat_y','ss_long_y']]
matrix_all_s.groupby(['site_id_x']).agg({'c':np.sum}).sort_values(by='c', ascending=False)[0:6]

matrix_all_s= gpd.GeoDataFrame(matrix_all_s,  geometry = [Point(xy) for xy in zip(matrix_all_s['ss_long_y'], matrix_all_s['ss_lat_y'])])
matrix_all_s.crs='epsg:4326'
matrix_all_s=matrix_all_s[matrix_all_s['site_id_x'].isin(['53','49','60','14','28','126'])]
matrix_all_s.groupby(['site_id_x']).apply(lambda x: x.to_file(fr"./res/{x.name}_video.json", index=False, driver="GeoJSON", encoding='utf-8'))
matrix_all2=matrix_all_s[matrix_all_s['site_id_2'].isin(['53','49','60','14','28','126'])]
matrix_all2=matrix_all2[['site_id_2','geometry']].drop_duplicates()
#matrix_all2.to_file(fr"./res/sites.json", index=False, driver="GeoJSON", encoding='utf-8')

## Объединяем с ТАЗами

sites_unic=pd.read_csv('sites_coordinate.csv', sep = ';', encoding='utf-8')

sites_unic[0:1]

sites_gpd = gpd.GeoDataFrame(sites_unic, \
                              geometry = [Point(xy) for xy in zip(sites_unic['ss_long'], sites_unic['ss_lat'])])
sites_gpd.crs='epsg:4326'

transport_zone=gpd.read_file(r'C:\Users\user\Documents\GIT\city\user_scripts\masha_k\Transnavigation\algoritm\557_zone.shp', encoding='utf-8')

site_taz=gpd.sjoin(sites_gpd, transport_zone, how='left', op='within', lsuffix='left', rsuffix='right')

site_taz=site_taz.rename(columns={'NO':'Транспортные районы'})

matrix=pd.read_csv('./res/matrix_Tolyatti.csv')

matrix_taz=matrix.merge(site_taz,on ='site_id', how='left')
matrix_taz=matrix_taz.merge(site_taz,left_on ='site_id_2',right_on ='site_id',how='left')

matrix_taz2=pd.DataFrame(matrix_taz.groupby(['Транспортные районы_x','Транспортные районы_y']).agg({'c':sum}).reset_index())

matrix_taz2.to_csv('matrix_taz_Tolyat.csv',index=False)

matrix_taz2['Транспортные районы_y'].unique()
